Match convolution widths, widen the normal-matrix accumulators, document both
Follow-up to the full-field memory work, from an audit of every float64
allocation in the package. The headline is that float64 is NOT what the
fitting run is made of: everything narrowable on the F770W fitting path
totals ~0.36 GB, against a peak that holds three full template sets at once,
all already float32. The one large win is on the detection path.

Narrowed (all of these are scipy promoting to the wider operand, so both
sides had to be matched or the promotion stayed):

  - Catalog._detect: Gaussian2DKernel(...).array is always float64 and was the
    sole promoter of a float32 detection mosaic. Casting it to the image dtype
    takes the 876 Mpx convolution from 25.4 GB peak to 14.5 GB. The FFT sums
    inside pocketfft so there is no accumulator to annotate; the error is
    ~log2(N) eps32 = 2e-6 relative, on a map thresholded at 2 sigma.
  - Templates.convolve_templates: dropped the .astype(float) on the stamp,
    whose result was cast straight back, and matched the kernel to the stamp
    width instead. Halves the FFT workspace of all 138,610 convolutions.
  - template_schemes.representative_psf: dropped dtype=float from the stack
    comprehension. Every consumer re-casts internally, so this is bit-exact,
    and it halves a 136 MB peak.
  - scene_fitter.build_normal: deleted `norms`, which was assigned and never
    read.

Deliberately NOT narrowed, though the audit proposed it:

  - PSFRegionMap stamps. Narrowing the stored cube would save 117 MB, but the
    class also holds the band PSF maps, whose finite stamp sums are the
    throughput metadata converting a fitted amplitude to a total flux. At
    float32 a unit-normalised stamp sums to 1 +/- 5e-9, which pushed a
    reported EE_BOX above 1 and failed the CLI test. The convolution width is
    matched at the point of use instead.
  - Bq/Bl in assemble_scene_system_AB. Overturned during adversarial verify:
    the += across overlapping anchors rounds before any reduction, so a wider
    accumulator on the final sum cannot recover it.

Widened, which is the inverse finding and a correctness fix. The normal
matrix was built with float(np.sum(cut_i * w_i * cut_i)) and no dtype=.
np.sum on float32 operands accumulates in float32, so ATA and ATb carried
about seven significant digits inside a float64 container, and the whitening,
Cholesky and spsolve chain downstream inherited that. Now accumulated in
float64 at scene_fitter.py (diagonal, RHS and off-diagonal) and fit.py
(_weighted_norm, atb, off-diagonal). Each entry sums ~1e4 terms over one
stamp footprint, so the wider accumulator is free.

Also reverted the float(t.flux) half of a local Scene.model_image edit while
keeping its dtype=np.float32 half. Under NEP 50 a Python float is a weak
scalar and would compute the product in float32 and round twice; t.flux is an
np.float64 from the solve, so leaving it strong forms an exact float64 product
and rounds once. The temporary is one stamp either way -- the buffer is where
the memory is.

docs/precision.md records the rule that separates the two cases, the NEP 50
and scipy-promotion mechanics, the table of what is float64 on purpose, and
why the segmentation map stays big-endian int32.

poetry run pytest: 356 passed.

+++ src/mophongo/scene.py
@@ -1530,13 +1530,18 @@ class Scene:
         if self.solution is None:
             raise RuntimeError("No solution available")
         bb = self.bbox
-        model_scene = np.zeros((bb[1] - bb[0] + 1, bb[3] - bb[2] + 1), dtype=float)
+        model_scene = np.zeros((bb[1] - bb[0] + 1, bb[3] - bb[2] + 1), dtype=np.float32)
         for t in self.templates:
             sl = t.slices_original
             sl_local_scene = (
                 slice(sl[0].start - bb[0], sl[0].stop - bb[0]),
                 slice(sl[1].start - bb[2], sl[1].stop - bb[2]),
             )
+            # t.flux is an np.float64 out of the solve, so this product forms in
+            # float64 and rounds once on the accumulate. Wrapping it in float()
+            # would make it a weak scalar (NEP 50), computing the product in
+            # float32 and rounding twice, for no saving: the temporary is one
+            # stamp either way, and the buffer above is where the memory is.
             model_scene[sl_local_scene] += t.flux * t.data[t.slices_cutout]
         return model_scene
